backend/requirements.txt
fastapi
uvicorn
sqlalchemy
google-genai
python-dotenv
pydantic
